# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/v1/login": {
            "post": {
                "tags": [
                    "Driver Auth"
                ],
                "summary": "Driver Login",
                "description": "Driver Login with Username and Password",
                "operationId": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user_name",
                        "in": "formData",
                        "description": "Driver Username",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Driver Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation returns driver token",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Errors",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Fields are invalid"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "username and password are not correct",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/v1/logout": {
            "get": {
                "tags": [
                    "Driver Auth"
                ],
                "summary": "Driver Logout",
                "operationId": "logout",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Driver token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver Logged out successfully",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/v1/tasks-bulk/createUnAssignedBulkOfTasks": {
            "post": {
                "tags": [
                    "bulk"
                ],
                "summary": "create un assigned bulk of tasks",
                "description": "Multiple status values can be provided with comma separated strings",
                "operationId": "createUnAssignedBulkOfTasks",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "tasks",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Pet"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "Pet": {
            "required": [
                "name",
                "photoUrls"
            ],
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "doggie"
                },
                "category": {
                    "description": ""
                },
                "photoUrls": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "xml": {
                        "name": "photoUrl",
                        "wrapped": true
                    }
                },
                "tags": {
                    "description": "",
                    "type": "array",
                    "items": {},
                    "xml": {
                        "name": "tag",
                        "wrapped": true
                    }
                },
                "status": {
                    "description": "",
                    "type": "string",
                    "enum": [
                        "available",
                        "pending",
                        "sold"
                    ]
                }
            },
            "type": "object",
            "xml": {
                "name": "Pet"
            }
        },
...
    },
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}